expire-guard-fullstack/server/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
pytesseract
numpy
